requirements.txt
Flask
Pillow
pywin32
